+++ index.js
@@ -9,63 +9,35 @@ const relationShipMiddlewareFactory = require('./middleware/relationship');
 const {createError, createErrorPromise} = require('./lib/error');
 const {RouteExposureHandler} = require('./lib/route');
 const _createReplyObject = require('./lib/reply/create-reply-object');
-
-// TODO: move to own file
-const _attachReply = (req, res, next, status, result, message) => {
-  res.__payload = {status, result, message};
-  next();
-  return Promise.resolve();
-};
-
-// TODO: move to own file, move _formatValidationError to the same file
-const _handleError = (next, err) => {
-  if (ERRORS.ValidationError(err))
-    return next(createError(400, _formatValidationError(err)));
-  else if (err.isCreatedError)
-    return next(err);
-  else
-    return next(createError(500, err));
-};
-
-// TODO: move to own file, move _handleError to the same file
-const _formatValidationError = (err) => {
-  return err.errors.map(error => {
-    return _.pick(error, ['type', 'path', 'value']);
-  });
-};
+const {createReplyObject, errorHandler, replyHandler} = require('./lib/reply');
 
 const _update = async (model, req, res, next, id, createInput) => {
-  const attachReply = _attachReply.bind(null, req, res, next);
-  const handleError = _handleError.bind(null, next);
-
   const attributes = model.getUpdateableAttributes().map(attribute => attribute.attribute);
   try {
     const instance = await model.findByPk(id);
     if (!instance) await createErrorPromise(404);
     await instance.update(createInput(req.body), {fields: attributes});
-    return attachReply(204);
+    return res.replyHandler(next, 204);
   } catch (err) {
-    return handleError(err);
+    return res.errorHandler(next, err);
   }
 };
 
 // TODO: this method is a mess, we need to clean it up
 const _updateRelation = async (createReplyObject, source, target, association, req, res, next, id, targetId, prepareBody) => {
-  const attachReply = _attachReply.bind(null, req, res, next);
-  const handleError = _handleError.bind(null, next);
   try {
     const sourceInstance = await source.findByPk(id);
     const update = _update.bind(null, target);
-    if (!sourceInstance) return attachReply(404, undefined, 'source not found.');
+    if (!sourceInstance) return res.replyHandler(next, 404, undefined, 'source not found.');
     const query =
       association.associationType === 'HasOne' || association.associationType === 'BelongsTo' ? undefined : {where: {id: targetId}};
     let targetInstance = await sourceInstance[association.accessors.get](query);
     if (!targetInstance) await createErrorPromise(404, 'target not found.');
-    if (targetInstance instanceof Array) // "many" relationsship
+    if (targetInstance instanceof Array) // "many" relationship
       targetInstance = targetInstance[0];
     await update(req, res, next, createReplyObject(targetInstance).id, prepareBody);
   } catch (err) {
-    handleError(err);
+    res.errorHandler(next, err);
   }
 };
 
@@ -196,6 +168,16 @@ module.exports = (models, opts) => {
 
     const auth = model.getAuthorizationMiddleWare.bind(model, null);
 
+    // attaching errorHandler and replyHandler to res
+    router.use((req, res, next) => {
+      // (req, res, next, status, result, message)
+      res.replyHandler = replyHandler.bind(null, req, res);
+      // (next, err)
+      res.errorHandler = errorHandler(ERRORS);
+
+      next();
+    });
+
     router.use((req, res, next) => {
       // resetting the state of the query builder before each request
       queryBuilder.reset();
@@ -210,19 +192,17 @@ module.exports = (models, opts) => {
     }
     if (routeExposureHandler.isRouteExposed('post', '/')) {
       router.post('/', auth('CREATE'), (req, res, next) => {
-        const attachReply = _attachReply.bind(null, req, res, next);
-        const handleError = _handleError.bind(null, next);
         const input = model.removeIllegalAttributes(req.body);
         model
           .create(input)
           .then(modelInstance => {
             if (routing.opts.filterReferenceAttributes) {
-              return attachReply(201, model.filterReferenceAttributesFromModelInstance(createReplyObject(modelInstance)));
+              return res.replyHandler(next, 201, model.filterReferenceAttributesFromModelInstance(createReplyObject(modelInstance)));
             } else {
-              return attachReply(201, createReplyObject(modelInstance));
+              return res.replyHandler(next, 201, createReplyObject(modelInstance));
             }
           }).catch(err => {
-            return handleError(err);
+            return res.errorHandler(next, err);
           });
       });
       openApiDocument.addOperationAndComponents(openApiBaseName, 'post', openApiHelper.createModelPathSpecification('post'));
@@ -230,9 +210,6 @@ module.exports = (models, opts) => {
 
     if (routeExposureHandler.isRouteExposed('post', '/bulk')) {
       router.post('/bulk', auth('CREATE'), async (req, res, next) => {
-        const attachReply = _attachReply.bind(null, req, res, next);
-        const handleError = _handleError.bind(null, next);
-
         const transaction = await model.transaction();
 
         try {
@@ -244,14 +221,14 @@ module.exports = (models, opts) => {
           }));
           await transaction.commit();
           if (routing.opts.filterReferenceAttributes) {
-            return attachReply(201,
+            return res.replyHandler(next, 201,
               createReplyObject(instances).map(instance => model.filterReferenceAttributesFromModelInstance(instance)));
           } else {
-            return attachReply(201, createReplyObject(instances));
+            return res.replyHandler(next, 201, createReplyObject(instances));
           }
         } catch (err) {
           await transaction.rollback();
-          return handleError(err);
+          return res.errorHandler(next, err);
         }
       });
       openApiDocument.addOperationAndComponents(`${openApiBaseName}/bulk`, 'post', openApiHelper.createBulkModelPathSpecification());
@@ -259,12 +236,10 @@ module.exports = (models, opts) => {
 
     if (routeExposureHandler.isRouteExposed('get', '/count')) {
       router.get('/count', auth('READ'), async (req, res, next) => {
-        const attachReply = _attachReply.bind(null, req, res, next);
-        const handleError = _handleError.bind(null, next);
         try {
-          return attachReply(200, await model.count(), `Count for ${model.name} obtained!`);
+          return res.replyHandler(next, 200, await model.count(), `Count for ${model.name} obtained!`);
         } catch (err) {
-          return handleError(err);
+          return res.errorHandler(next, err);
         }
       });
       openApiDocument.addOperationAndComponents(`${openApiBaseName}/count`, 'get', openApiHelper.createCountModelPathSpecification());
@@ -272,8 +247,6 @@ module.exports = (models, opts) => {
 
     if (routeExposureHandler.isRouteExposed('get', '/')) {
       router.get('/', auth('READ'), async (req, res, next) => {
-        const attachReply = _attachReply.bind(null, req, res, next);
-        const handleError = _handleError.bind(null, next);
         try {
           const results = await model.findAll(
             queryBuilder
@@ -281,9 +254,9 @@ module.exports = (models, opts) => {
               .prepare()
               .query
           );
-          return attachReply(200, createReplyObject(results));
+          return res.replyHandler(next, 200, createReplyObject(results));
         } catch (err) {
-          return handleError(err);
+          return res.errorHandler(next, err);
         }
       });
       openApiDocument.addOperationAndComponents(openApiBaseName, 'get', openApiHelper.createModelPathSpecification('get'));
@@ -291,8 +264,6 @@ module.exports = (models, opts) => {
 
     if (routeExposureHandler.isRouteExposed('get', '/search') || routeExposureHandler.isRouteExposed('post', '/search')) {
       router.post('/search', auth('SEARCH'), async (req, res, next) => {
-        const attachReply = _attachReply.bind(null, req, res, next);
-        const handleError = _handleError.bind(null, next);
         try {
           const results = await model.findAll(
             queryBuilder
@@ -311,12 +282,12 @@ module.exports = (models, opts) => {
               .query
           ));
           if (results.length === 0) {
-            return attachReply(204);
+            return res.replyHandler(next, 204);
           } else {
-            return attachReply(200, createReplyObject(results));
+            return res.replyHandler(next, 200, createReplyObject(results));
           }
         } catch (err) {
-          return handleError(err);
+          return res.errorHandler(next, err);
         }
       });
       openApiDocument.addOperationAndComponents(`${openApiBaseName}/search`, 'post', openApiHelper.createSearchModelPathSpecification());
@@ -325,15 +296,12 @@ module.exports = (models, opts) => {
     if (routeExposureHandler.isRouteExposed('get', '/:id')) {
       router.get(`/:id${idRegex}`, auth('READ'), (req, res, next) => {
         const id = req.params.id;
-        const attachReply = _attachReply.bind(null, req, res, next);
-        const handleError = _handleError.bind(null, next);
-
         const attributes = req.query.a ? req.query.a.split('|') : undefined;
         model.findOne({where: {id}, attributes}).then(modelInstance => {
           if (!modelInstance) return createErrorPromise(404, 'entity not found.');
-          return attachReply(200, createReplyObject(modelInstance));
+          return res.replyHandler(next, 200, createReplyObject(modelInstance));
         }).catch(err => {
-          return handleError(err);
+          return res.errorHandler(next, err);
         });
       });
       openApiDocument.addOperationAndComponents(`${openApiBaseName}/{id}`, 'get', openApiHelper.createInstancePathSpecification('get'));
@@ -359,15 +327,13 @@ module.exports = (models, opts) => {
 
     if (routeExposureHandler.isRouteExposed('delete', '/:id')) {
       router.delete(`/:id${idRegex}`, auth('DELETE'), async (req, res, next) => {
-        const attachReply = _attachReply.bind(null, req, res, next);
-        const handleError = _handleError.bind(null, next);
         try {
           const instance = await model.findByPk(req.params.id);
           if (!instance) await createErrorPromise(404);
           await instance.destroy();
-          return attachReply(204);
+          return res.replyHandler(next, 204);
         } catch (err) {
-          return handleError(err);
+          return res.errorHandler(next, err);
         }
       });
       openApiDocument.addOperationAndComponents(
@@ -383,31 +349,26 @@ module.exports = (models, opts) => {
       const auth = target.getAuthorizationMiddleWare.bind(target, source);
       // TODO: move into own file (maybe with update)
       const unlinkRelations = (req, res, next, setterFunctionName) => {
-        const attachReply = _attachReply.bind(null, req, res, next);
-        const handleError = _handleError.bind(null, next);
-
         source.findByPk(req.params.id).then(sourceInstance => {
           if (!sourceInstance) return createErrorPromise(404, 'source not found.');
           return sourceInstance[setterFunctionName](null).then(_ => {
-            return attachReply(204);
+            return res.replyHandler(next, 204);
           });
         }).catch(err => {
-          return handleError(err);
+          return res.errorHandler(next, err);
         });
       };
       // TODO: move into own file (maybe with update)
       const relationshipGet = (postProcess) => {
         return (req, res, next) => {
-          const attachReply = _attachReply.bind(null, req, res, next);
-          const handleError = _handleError.bind(null, next);
           source.findByPk(req.params.id).then(sourceInstance => {
             if (!sourceInstance) return createErrorPromise(404, 'source not found.');
             return sourceInstance[association.accessors.get]().then(targetInstance => {
               if (!targetInstance) return createErrorPromise(404, 'target not found.');
-              return attachReply(200, postProcess(req, targetInstance));
+              return res.replyHandler(next, 200, postProcess(req, targetInstance));
             });
           }).catch(err => {
-            return handleError(err);
+            return res.errorHandler(next, err);
           });
         };
       };
@@ -439,8 +400,6 @@ module.exports = (models, opts) => {
           }
           if (routeExposureHandler.isRouteExposed('post', baseTargetRouteOpt)) {
             router.post(`/:id${idRegex}/${targetRoute}`, auth('CREATE'), async (req, res, next) => {
-              const attachReply = _attachReply.bind(null, req, res, next);
-              const handleError = _handleError.bind(null, next);
               try {
                 const sourceInstance = await source.findByPk(req.params.id);
                 if (!sourceInstance) return await createErrorPromise(404, 'source not found.');
@@ -449,10 +408,10 @@ module.exports = (models, opts) => {
                 if (association.associationType === 'BelongsTo') {
                   if (instance instanceof source) { // 4.x.x
                     return sourceInstance[association.accessors.get]().then(createdTargetInstance => {
-                      return attachReply(201, createReplyObject(createdTargetInstance));
+                      return res.replyHandler(next, 201, createReplyObject(createdTargetInstance));
                     });
                   } else if (instance instanceof target) { // 5.x.x
-                    return attachReply(201, createReplyObject(instance));
+                    return res.replyHandler(next, 201, createReplyObject(instance));
                   } else {
                     createErrorPromise(
                       500,
@@ -460,10 +419,10 @@ module.exports = (models, opts) => {
                     );
                   }
                 } else {
-                  return attachReply(201, createReplyObject(instance));
+                  return res.replyHandler(next, 201, createReplyObject(instance));
                 }
               } catch (err) {
-                handleError(err);
+                res.errorHandler(next, err);
               }
             });
             openApiDocument.addOperationAndComponents(
@@ -513,14 +472,12 @@ module.exports = (models, opts) => {
           }
           if (routeExposureHandler.isRouteExposed('get', `${baseTargetRouteOpt}/count`)) {
             router.get(`/:id${idRegex}/${targetRoute}/count`, auth('READ'), async (req, res, next) => {
-              const attachReply = _attachReply.bind(null, req, res, next);
-              const handleError = _handleError.bind(null, next);
               try {
-                return attachReply(200,
+                return res.replyHandler(next, 200,
                   await source.getAssociationCount(association, req.params.id),
                   `Count for ${model.name} obtained!`);
               } catch (err) {
-                return handleError(err);
+                return res.errorHandler(next, err);
               }
             });
             openApiDocument.addOperationAndComponents(
@@ -530,8 +487,6 @@ module.exports = (models, opts) => {
 
           if (routeExposureHandler.isRouteExposed('post', `${baseTargetRouteOpt}/search`)) {
             router.post(`/:id${idRegex}/${targetRoute}/search`, auth('SEARCH'), async (req, res, next) => {
-              const attachReply = _attachReply.bind(null, req, res, next);
-              const handleError = _handleError.bind(null, next);
               try {
                 const searchQuery = queryBuilder
                   .create(req.body)
@@ -541,12 +496,12 @@ module.exports = (models, opts) => {
                 const [searchOptions, results] = await _searchBySourceIdAndTargetQuery(association, req.params.id, searchQuery);
                 res.set('X-Total-Count', await source.getAssociationCount(association, req.params.id, searchOptions));
                 if (results.length === 0) {
-                  return attachReply(204);
+                  return res.replyHandler(next, 204);
                 } else {
-                  return attachReply(200, createReplyObject(results));
+                  return res.replyHandler(next, 200, createReplyObject(results));
                 }
               } catch (err) {
-                return handleError(err);
+                return res.errorHandler(next, err);
               }
             });
             openApiDocument.addOperationAndComponents(
@@ -556,8 +511,6 @@ module.exports = (models, opts) => {
 
           if (routeExposureHandler.isRouteExposed('get', instanceTargetRouteOpt)) {
             router.get(`/:id${idRegex}/${targetRoute}/:targetId${idRegex}`, auth('READ'), (req, res, next) => {
-              const attachReply = _attachReply.bind(null, req, res, next);
-              const handleError = _handleError.bind(null, next);
               source.findByPk(req.params.id).then(async sourceInstance => {
                 if (!sourceInstance) return createErrorPromise(404, 'source not found.');
 
@@ -567,10 +520,10 @@ module.exports = (models, opts) => {
                 if (instances[0] instanceof source) { // 4.x.x
                   return sourceInstance[association.accessors.get]({where: {id: {$eq: req.params.targetId}}}).spread(targetInstance => {
                     if (!targetInstance) return createErrorPromise(404, 'target not found.');
-                    return attachReply(200, _filterAttributes(req.query.a, createReplyObject(targetInstance)));
+                    return res.replyHandler(next, 200, _filterAttributes(req.query.a, createReplyObject(targetInstance)));
                   });
                 } else if (instances[0] instanceof target) { // 5.x.x
-                  return attachReply(200, _filterAttributes(req.query.a, createReplyObject(instances[0])));
+                  return res.replyHandler(next, 200, _filterAttributes(req.query.a, createReplyObject(instances[0])));
                 } else {
                   return createErrorPromise(
                     500,
@@ -578,7 +531,7 @@ module.exports = (models, opts) => {
                   );
                 }
               }).catch(err => {
-                return handleError(err);
+                return res.errorHandler(next, err);
               });
             });
             openApiDocument.addOperationAndComponents(
@@ -588,15 +541,13 @@ module.exports = (models, opts) => {
 
           if (routeExposureHandler.isRouteExposed('post', baseTargetRouteOpt)) {
             router.post(`/:id${idRegex}/${targetRoute}`, auth('CREATE'), (req, res, next) => {
-              const attachReply = _attachReply.bind(null, req, res, next);
-              const handleError = _handleError.bind(null, next);
               source.findByPk(req.params.id).then(sourceInstance => {
                 if (!sourceInstance) return createErrorPromise(404, 'source not found.');
                 return sourceInstance[association.accessors.create](target.removeIllegalAttributes(req.body));
               }).then(instance => {
-                return attachReply(201, createReplyObject(instance));
+                return res.replyHandler(next, 201, createReplyObject(instance));
               }).catch(err => {
-                return handleError(err);
+                return res.errorHandler(next, err);
               });
             });
             openApiDocument.addOperationAndComponents(
@@ -606,8 +557,6 @@ module.exports = (models, opts) => {
           if (association.associationType === 'BelongsToMany') {
             if (routeExposureHandler.isRouteExposed('post', `${instanceTargetRouteOpt}/link`)) {
               router.post(`/:id${idRegex}/${targetRoute}/:targetId${idRegex}/link`, auth('ASSOCIATE'), async (req, res, next) => {
-                const attachReply = _attachReply.bind(null, req, res, next);
-                const handleError = _handleError.bind(null, next);
                 try {
                   const sourceInstance = await source.findByPk(req.params.id);
                   if (!sourceInstance) return createErrorPromise(404, 'source not found.');
@@ -616,9 +565,9 @@ module.exports = (models, opts) => {
 
                   await sourceInstance[association.accessors.add](targetInstance);
 
-                  return attachReply(204);
+                  return res.replyHandler(next, 204);
                 } catch (err) {
-                  return handleError(err);
+                  return res.errorHandler(next, err);
                 }
               });
               openApiDocument.addOperationAndComponents(
@@ -628,8 +577,6 @@ module.exports = (models, opts) => {
 
             if (routeExposureHandler.isRouteExposed('delete', `${instanceTargetRouteOpt}/unlink`)) {
               router.delete(`/:id${idRegex}/${targetRoute}/:targetId${idRegex}/unlink`, auth('ASSOCIATE'), async (req, res, next) => {
-                const attachReply = _attachReply.bind(null, req, res, next);
-                const handleError = _handleError.bind(null, next);
                 try {
                   const sourceInstance = await source.findByPk(req.params.id);
                   if (!sourceInstance) return createErrorPromise(404, 'source not found.');
@@ -639,9 +586,9 @@ module.exports = (models, opts) => {
 
                   await sourceInstance[association.accessors.remove](targetInstance);
 
-                  return attachReply(204);
+                  return res.replyHandler(next, 204);
                 } catch (err) {
-                  return handleError(err);
+                  return res.errorHandler(next, err);
                 }
               });
               openApiDocument.addOperationAndComponents(
@@ -683,9 +630,6 @@ module.exports = (models, opts) => {
           }
           if (routeExposureHandler.isRouteExposed('delete', instanceTargetRouteOpt)) {
             router.delete(`/:id${idRegex}/${targetRoute}/:targetId${idRegex}`, auth('DELETE'), (req, res, next) => {
-              const attachReply = _attachReply.bind(null, req, res, next);
-              const handleError = _handleError.bind(null, next);
-
               source.findByPk(req.params.id).then(sourceInstance => {
                 if (!sourceInstance) return createErrorPromise(404, 'source not found.');
                 return sourceInstance[association.accessors.get]({where: {id: req.params.targetId}}).then(targetInstances => {
@@ -693,10 +637,10 @@ module.exports = (models, opts) => {
                   if (!targetInstance) return createErrorPromise(404, 'target not found.');
                   return sourceInstance[association.accessors.remove](targetInstance);
                 }).then(() => {
-                  return attachReply(204);
+                  return res.replyHandler(next, 204);
                 });
               }).catch(err => {
-                return handleError(err);
+                return res.errorHandler(next, err);
               });
             });
             openApiDocument.addOperationAndComponents(